Fix fixed-pitch font not adjusting size
Setting the `doom-font` variable in doom emacs will configure both the
`default` and `fixed-pitch` "faces" in emacs with a size, but the
font size scaler in emacs only affects the `default` face, leaving
`fixed-pitch` unchanged.

If `fixed-pitch` is instead set to be a fraction it will always be sized
relative to the `default` face, and as such it will be scaled
appropriately.